Consolidate deeplink handling into dedicated module
