+++ src/lib/deeplink/handlers/auth.ts
@@ -1,0 +1,71 @@
+import * as Sentry from '@sentry/electron/main';
+import { z } from 'zod';
+import { sendIpcEventToRenderer } from 'src/ipc-utils';
+import wpcomFactory from 'src/lib/wpcom-factory';
+import wpcomXhrRequest from 'src/lib/wpcom-xhr-request-factory';
+import { updateAppdata } from 'src/storage/user-data';
+
+const authTokenSchema = z.object( {
+	accessToken: z.string(),
+	expiresIn: z.number(),
+	expirationTime: z.number(),
+	id: z.number(),
+	email: z.string(),
+	displayName: z.string().default( '' ),
+} );
+
+const meResponseSchema = z.object( {
+	ID: z.number(),
+	email: z.string(),
+	display_name: z.string(),
+} );
+
+export type StoredToken = z.infer< typeof authTokenSchema >;
+
+async function handleAuthCallback( hash: string ): Promise< StoredToken > {
+	const params = new URLSearchParams( hash.substring( 1 ) );
+	const error = params.get( 'error' );
+
+	if ( error ) {
+		// Close the browser if code found or error
+		throw new Error( error );
+	}
+
+	const accessToken = params.get( 'access_token' ) ?? '';
+	const expiresIn = parseInt( params.get( 'expires_in' ) ?? '0' );
+
+	if ( isNaN( expiresIn ) || expiresIn === 0 || ! accessToken ) {
+		throw new Error( 'Error while getting token' );
+	}
+	const rawResponse = await wpcomFactory( accessToken, wpcomXhrRequest ).req.get(
+		'/me?fields=ID,email,display_name'
+	);
+
+	const response = meResponseSchema.parse( rawResponse );
+
+	return authTokenSchema.parse( {
+		expiresIn,
+		expirationTime: new Date().getTime() + expiresIn * 1000,
+		accessToken,
+		id: response.ID,
+		email: response.email,
+		displayName: response.display_name,
+	} );
+}
+
+/**
+ * Handles the OAuth authentication deeplink callback.
+ * This function is called when the user completes authentication on WordPress.com
+ * and is redirected back to the app via wpcom-local-dev://auth
+ */
+export async function handleAuthDeeplink( urlObject: URL ): Promise< void > {
+	const { hash } = urlObject;
+	try {
+		const authResult = await handleAuthCallback( hash );
+		await updateAppdata( { authToken: authResult } );
+		void sendIpcEventToRenderer( 'auth-updated', { token: authResult } );
+	} catch ( error ) {
+		Sentry.captureException( error );
+		void sendIpcEventToRenderer( 'auth-updated', { error } );
+	}
+}
