+++ src/lib/deeplink/index.ts
@@ -1,0 +1,3 @@
+export { handleDeeplink } from './deeplink-handler';
+export { handleAuthDeeplink } from './handlers/auth';
+export { handleSyncConnectSiteDeeplink } from './handlers/sync-connect-site';
