+++ src/lib/oauth.ts
@@ -1,12 +1,8 @@
-import * as Sentry from '@sentry/electron/main';
 import { z } from 'zod';
 import { CLIENT_ID } from 'common/constants';
 import { SupportedLocale } from 'common/lib/locale';
 import { getAuthenticationUrl } from 'common/lib/oauth';
-import { sendIpcEventToRenderer } from 'src/ipc-utils';
-import wpcomFactory from 'src/lib/wpcom-factory';
-import wpcomXhrRequest from 'src/lib/wpcom-xhr-request-factory';
-import { loadUserData, updateAppdata } from 'src/storage/user-data';
+import { loadUserData } from 'src/storage/user-data';
 
 const authTokenSchema = z.object( {
 	accessToken: z.string(),
@@ -15,12 +11,6 @@ const authTokenSchema = z.object( {
 	id: z.number(),
 	email: z.string(),
 	displayName: z.string().default( '' ),
-} );
-
-const meResponseSchema = z.object( {
-	ID: z.number(),
-	email: z.string(),
-	display_name: z.string(),
 } );
 
 export type StoredToken = z.infer< typeof authTokenSchema >;
@@ -51,52 +41,4 @@ export async function getAuthenticationToken(): Promise< StoredToken | null > {
 export async function isAuthenticated(): Promise< boolean > {
 	const token = await getAuthenticationToken();
 	return !! token;
-}
-
-async function handleAuthCallback( hash: string ): Promise< StoredToken > {
-	const params = new URLSearchParams( hash.substring( 1 ) );
-	const error = params.get( 'error' );
-
-	if ( error ) {
-		// Close the browser if code found or error
-		throw new Error( error );
-	}
-
-	const accessToken = params.get( 'access_token' ) ?? '';
-	const expiresIn = parseInt( params.get( 'expires_in' ) ?? '0' );
-
-	if ( isNaN( expiresIn ) || expiresIn === 0 || ! accessToken ) {
-		throw new Error( 'Error while getting token' );
-	}
-	const rawResponse = await wpcomFactory( accessToken, wpcomXhrRequest ).req.get(
-		'/me?fields=ID,email,display_name'
-	);
-
-	const response = meResponseSchema.parse( rawResponse );
-
-	return authTokenSchema.parse( {
-		expiresIn,
-		expirationTime: new Date().getTime() + expiresIn * 1000,
-		accessToken,
-		id: response.ID,
-		email: response.email,
-		displayName: response.display_name,
-	} );
-}
-
-/**
- * Handles the OAuth authentication deeplink callback.
- * This function is called when the user completes authentication on WordPress.com
- * and is redirected back to the app via wpcom-local-dev://auth
- */
-export async function handleAuthDeeplink( urlObject: URL ): Promise< void > {
-	const { hash } = urlObject;
-	try {
-		const authResult = await handleAuthCallback( hash );
-		await updateAppdata( { authToken: authResult } );
-		void sendIpcEventToRenderer( 'auth-updated', { token: authResult } );
-	} catch ( error ) {
-		Sentry.captureException( error );
-		void sendIpcEventToRenderer( 'auth-updated', { error } );
-	}
 }
